platform/platform-core/fastapi/pyproject.toml
[project]
name = "platform-core-fastapi"
version = "0.1.0"
description = "Platform Core - FastAPI Runtime"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = ">=0.115.0"
uvicorn = ">=0.34.0"
pyyaml = ">=6.0"
pydantic = ">=2.0"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["platform_core_fastapi"]

[project.scripts]
platform-fastapi = "platform_core_fastapi.bootstrap:main"
